requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
pycocotools>=2.0.2
Pillow>=8.0.0
numpy>=1.19.0